+++ react/__mocks__/vtex.delivery-promise-components/DeliveryPromiseContext.js
@@ -1,1 +1,2 @@
 export const useDeliveryPromiseState = () => ({})
+export const useDeliveryPromiseDispatch = () => () => {}