+++ react/__mocks__/vtex.delivery-promise-components/PickupModalPresentational.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const PickupModalPresentational = () => (
+  <div data-testid="pickup-modal-presentational" />
+)
+
+export default PickupModalPresentational